# openapi.json
{"openapi":"3.0.1","info":{"title":"Pet-Friendly API","description":"Backend REST API for the Pet-Friendly mobile application — places, reviews, favorites, notifications.","contact":{"name":"Pet-Friendly","email":"contact@petfriendly.app"},"license":{"name":"Proprietary"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Local dev"},{"url":"https://api.petfriendly.app","description":"Production"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Notifications","description":"In-app notifications for the current user"},{"name":"Users","description":"Current user profile and stats"},{"name":"Favorites","description":"Manage the current user's favorite places"},{"name":"Reviews","description":"Reviews left on places"},{"name":"Auth","description":"Registration, login, refresh"},{"name":"Places","description":"Browse, create, update and search pet-friendly places"}],"paths":{"/api/v1/users/me":{"get":{"tags":["Users"],"summary":"Get current user profile","description":"Returns the authenticated user's profile plus aggregated stats: reviewsWritten, favoritesCount, placesAdded.","operationId":"me","responses":{"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"200":{"description":"Profile and stats","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}},"put":{"tags":["Users"],"summary":"Update current user profile","description":"Updates name and pets list.","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"400":{"description":"Invalid payload","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"200":{"description":"Profile updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/places/{id}":{"get":{"tags":["Places"],"summary":"Get place by id","description":"Public endpoint.","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Place not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"200":{"description":"Place found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}}},"security":[]},"put":{"tags":["Places"],"summary":"Update place","description":"Owner-only.","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlaceRequest"}}},"required":true},"responses":{"404":{"description":"Place not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"403":{"description":"Not the owner","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"200":{"description":"Place updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}}}},"delete":{"tags":["Places"],"summary":"Delete place","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Deleted"},"404":{"description":"Place not found"},"401":{"description":"Missing or invalid JWT"}}}},"/api/v1/users/favorites/{placeId}":{"post":{"tags":["Favorites"],"summary":"Toggle favorite","description":"Adds the place to favorites if not present, removes it otherwise.","operationId":"toggle","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Place not found"},"401":{"description":"Missing or invalid JWT"},"200":{"description":"Toggle applied"}}},"delete":{"tags":["Favorites"],"summary":"Remove favorite","description":"Idempotent — succeeds even if the place was not a favorite.","operationId":"remove","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"401":{"description":"Missing or invalid JWT"},"204":{"description":"Removed (or already absent)"}}}},"/api/v1/places":{"get":{"tags":["Places"],"summary":"List places","description":"Paginated list of all places. Public endpoint.","operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponsePlaceResponse"}}}}},"security":[]},"post":{"tags":["Places"],"summary":"Create place","description":"The authenticated user becomes the owner.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlaceRequest"}}},"required":true},"responses":{"400":{"description":"Invalid payload","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}},"201":{"description":"Place created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceResponse"}}}}}}},"/api/v1/places/{placeId}/reviews":{"get":{"tags":["Reviews"],"summary":"List reviews of a place","description":"Paginated, ordered by recency. Public endpoint.","operationId":"getByPlace","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseReviewResponse"}}}}},"security":[]},"post":{"tags":["Reviews"],"summary":"Create review","description":"Authenticated. The author is the current user.","operationId":"create_1","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewRequest"}}},"required":true},"responses":{"400":{"description":"Invalid payload (e.g. rating out of range)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"404":{"description":"Place not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"201":{"description":"Review created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Auth"],"summary":"Register a new user","description":"Creates an account and returns JWT access + refresh tokens.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"security":[]}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Refresh JWT","description":"Exchanges a valid refresh token for a new access token. The old refresh token is revoked (rotation).","operationId":"refresh","parameters":[{"name":"refreshToken","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"security":[]}},"/api/v1/auth/logout":{"post":{"tags":["Auth"],"summary":"Logout","description":"Revokes the provided refresh token so it cannot be re-used.","operationId":"logout","parameters":[{"name":"refreshToken","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}},"security":[]}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Authenticates an existing user and returns JWT tokens.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"security":[]}},"/api/v1/notifications/{id}/read":{"patch":{"tags":["Notifications"],"summary":"Mark as read","operationId":"markRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Marked as read"},"404":{"description":"Notification not found or not owned"},"401":{"description":"Missing or invalid JWT"}}}},"/api/v1/users/favorites":{"get":{"tags":["Favorites"],"summary":"List favorites","description":"Returns the current user's favorite places.","operationId":"getFavorites","responses":{"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaceResponse"}}}}}}}},"/api/v1/places/search":{"get":{"tags":["Places"],"summary":"Search places","description":"Filter by free-text query, type, accepted animals and geographic radius. When lat/lng are provided, uses PostGIS ST_DWithin (radius in km), otherwise falls back to LIKE search. Public endpoint.","operationId":"search","parameters":[{"name":"q","in":"query","description":"Free-text query (matches place name)","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","description":"Place type filter","required":false,"schema":{"type":"string","enum":["RESTAURANT","CAFE","HOTEL","PARC","COMMERCE"]}},{"name":"animals","in":"query","description":"Accepted animals filter (multi-value)","required":false,"schema":{"type":"array","items":{"type":"string","enum":["DOG","CAT","OTHER"]}}},{"name":"lat","in":"query","description":"Latitude in decimal degrees","required":false,"schema":{"type":"number","format":"double"}},{"name":"lng","in":"query","description":"Longitude in decimal degrees","required":false,"schema":{"type":"number","format":"double"}},{"name":"radius","in":"query","description":"Search radius in kilometers (default 5)","required":false,"schema":{"type":"number","format":"double","default":5.0}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponsePlaceResponse"}}}}},"security":[]}},"/api/v1/notifications":{"get":{"tags":["Notifications"],"summary":"List notifications","description":"Returns all notifications for the current user, most recent first.","operationId":"list_1","responses":{"401":{"description":"Missing or invalid JWT","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}},"delete":{"tags":["Notifications"],"summary":"Clear all notifications","description":"Deletes every notification of the current user.","operationId":"clearAll","responses":{"204":{"description":"Cleared"}}}},"/api/v1/reviews/{reviewId}":{"delete":{"tags":["Reviews"],"summary":"Delete review","description":"Author-only.","operationId":"delete","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Review not found"},"204":{"description":"Deleted"},"401":{"description":"Missing or invalid JWT"},"403":{"description":"Not the author"}}}},"/api/v1/notifications/{id}":{"delete":{"tags":["Notifications"],"summary":"Delete notification","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Deleted"},"404":{"description":"Notification not found or not owned"},"401":{"description":"Missing or invalid JWT"}}}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"name":{"type":"string","example":"Alice Dupont"},"pets":{"type":"array","example":["Rex"],"items":{"type":"string","example":"[\"Rex\"]"}}},"description":"Profile fields the current user can update"},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"name":{"type":"string"},"avatarUrl":{"type":"string"},"pets":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"}}},"CreatePlaceRequest":{"required":["address","latitude","longitude","name","type"],"type":"object","properties":{"name":{"type":"string","example":"Café des Toutous"},"type":{"type":"string","example":"CAFE","enum":["RESTAURANT","CAFE","HOTEL","PARC","COMMERCE"]},"address":{"type":"string","example":"12 rue de Rivoli, 75004 Paris"},"latitude":{"type":"number","description":"Latitude (decimal degrees, WGS84)","format":"double","example":48.8566},"longitude":{"type":"number","description":"Longitude (decimal degrees, WGS84)","format":"double","example":2.3522},"animals":{"type":"array","example":["DOG","CAT"],"items":{"type":"string","example":"[\"DOG\",\"CAT\"]","enum":["DOG","CAT","OTHER"]}},"description":{"type":"string","example":"Welcoming café with water bowls and a quiet terrace."},"hours":{"type":"object","additionalProperties":{"type":"string","description":"Opening hours per day (FR keys)","example":"{\"lundi\":\"09h-19h\",\"mardi\":\"09h-19h\"}"},"description":"Opening hours per day (FR keys)","example":{"lundi":"09h-19h","mardi":"09h-19h"}}},"description":"Payload to create or update a place"},"CoordinatesResponse":{"type":"object","properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"PlaceResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"type":{"type":"string","enum":["RESTAURANT","CAFE","HOTEL","PARC","COMMERCE"]},"address":{"type":"string"},"coordinates":{"$ref":"#/components/schemas/CoordinatesResponse"},"rating":{"type":"number","format":"double"},"reviewCount":{"type":"integer","format":"int32"},"animals":{"type":"array","items":{"type":"string","enum":["DOG","CAT","OTHER"]}},"imageUrl":{"type":"string"},"galleryUrls":{"type":"array","items":{"type":"string"}},"description":{"type":"string"},"hours":{"type":"object","additionalProperties":{"type":"string"}}}},"CreateReviewRequest":{"required":["rating"],"type":"object","properties":{"rating":{"maximum":5.0,"exclusiveMaximum":false,"minimum":1.0,"exclusiveMinimum":false,"type":"number","format":"double","example":4.5},"text":{"type":"string","example":"Mon chien a adoré, accueil au top."}},"description":"Payload to create a review on a place"},"ReviewResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"placeId":{"type":"string","format":"uuid"},"authorName":{"type":"string"},"authorAvatarUrl":{"type":"string"},"rating":{"type":"number","format":"double"},"text":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","name","password"],"type":"object","properties":{"email":{"type":"string","example":"alice@petfriendly.fr"},"password":{"maxLength":2147483647,"minLength":6,"type":"string","example":"secret123"},"name":{"type":"string","example":"Alice"},"pets":{"type":"array","description":"Optional list of pet names","example":["Rex","Whiskers"],"items":{"type":"string","description":"Optional list of pet names","example":"[\"Rex\",\"Whiskers\"]"}}},"description":"Payload to register a new user"},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserResponse"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","example":"user@petfriendly.fr"},"password":{"type":"string","example":"user123"}},"description":"Login credentials"},"PageResponsePlaceResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/PlaceResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}},"PageResponseReviewResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ReviewResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}},"NotificationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["NEW_PLACE","NEW_REVIEW","FAVORITE_SALE","SYSTEM","REMINDER"]},"title":{"type":"string"},"body":{"type":"string"},"read":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT access token obtained via POST /api/v1/auth/login","scheme":"bearer","bearerFormat":"JWT"}}}}